+++ modules/Scheduling/includes/RequestsReport.php
@@ -8,19 +8,38 @@
  * @subpackage Scheduling
  */
 
+// @since 7.8 Add Include Inactive Students checkbox.
+DrawHeader(
+	CheckBoxOnclick(
+		'include_inactive',
+		_( 'Include Inactive Students' )
+	)
+);
+
+$is_include_inactive = isset( $_REQUEST['include_inactive'] ) && $_REQUEST['include_inactive'] === 'Y';
+
+$where_active_sql = '';
+
+if ( ! $is_include_inactive )
+{
+	$where_active_sql = " AND '" . DBDate() . "'>=START_DATE
+	AND ('" . DBDate() . "'<=END_DATE OR END_DATE IS NULL)
+	AND MARKING_PERIOD_ID IN (" . GetAllMP( 'QTR', UserMP() ) . ")";
+}
+
 $count_RET = DBGet( "SELECT cs.TITLE as SUBJECT_TITLE,c.TITLE as COURSE_TITLE,sr.COURSE_ID,COUNT(*) AS COUNT,
 	(SELECT sum(TOTAL_SEATS)
 		FROM COURSE_PERIODS cp
 		WHERE cp.COURSE_ID=sr.COURSE_ID) AS SEATS,
 	(SELECT count(STUDENT_ID)
 		FROM SCHEDULE s
-		WHERE s.COURSE_ID=sr.COURSE_ID) AS STUDENTS
+		WHERE s.COURSE_ID=sr.COURSE_ID" . $where_active_sql . ") AS STUDENTS
 	FROM SCHEDULE_REQUESTS sr,COURSES c,COURSE_SUBJECTS cs
 	WHERE cs.SUBJECT_ID=c.SUBJECT_ID
 	AND sr.COURSE_ID=c.COURSE_ID
 	AND sr.SYEAR='" . UserSyear() . "'
 	AND sr.SCHOOL_ID='" . UserSchool() . "'
-	GROUP BY sr.COURSE_ID,cs.TITLE,c.TITLE" );
+	GROUP BY cs.SORT_ORDER,cs.TITLE,sr.COURSE_ID,c.TITLE" );
 
 $columns = array(
 	'SUBJECT_TITLE' => _( 'Subject' ),